Make listening port and host configurable via environment variables
It is generally a good idea to only listen on localhost per default.
When running inside of a container, however, we have to listen on
the public interface of the container. Otherwise, the service
is not accessible from the host.

This commit introduces environment variables MMT_LISTEN_HOST and
MMT_LISTEN_PORT to allow configuration of listen host and port,
respectively. The default is still 127.0.0.1:3000.